+++ src/inspect_ai/_view/www/src/app/types.ts
@@ -117,8 +117,6 @@ export interface LogState {
   filter: string;
   filterError?: FilterError;
 
-  epoch: string;
-  sort: string;
   selectedScores?: ScoreLabel[];
   scores?: ScoreLabel[];